requirements.txt
streamlit
google-generativeai
gspread
gspread-dataframe
pandas
biopython
pyperclip
markdown2
pywin32